Drink structure revamp + first implementation of drag
Deleted the ex Drink enum being changed with the Drink structure and the new DrinkType enum both collected in the Drinks folder. DrinkModel contains the new enum DrinkType and the new Drink struct, instead the DrinkModelView contains all the drinks and an array of all the usable drink. Deleted the useless class ToolBarView